+++ src/commands/version/get.ts
@@ -1,7 +1,9 @@
 import { Command, flags } from '@oclif/command'
+import chalk from 'chalk'
 
-import getAppCurrentData from '@helpers/getAppCurrentData'
-import getNextVersion from '@helpers/getNextVersion'
+import log from '@helpers/log-messages'
+import getAppCurrentData from '@helpers/get-app-current-data'
+import getNextVersion from '@helpers/get-next-version'
 
 export default class VersionGet extends Command {
   static description = 'show the current version of the app'
@@ -16,15 +18,18 @@ export default class VersionGet extends Command {
     const { flags: flag } = this.parse(VersionGet)
     const nextVersion = await getNextVersion()
     const currentVersion = getAppCurrentData(flag.folderPath)
-    let returnText = `The current version is ${currentVersion?.version}`
+    let returnText = `The current version is ${chalk.inverse(' ' + currentVersion?.version + ' ')}`
 
     if (Object.keys(nextVersion).includes('name')) {
-      returnText = `You are in a ${nextVersion.type} branch with a version number ${nextVersion.name}. ${returnText}`
+      returnText = `You are in a ${chalk.underline(nextVersion.type)} branch with a version number ${chalk.underline(
+        nextVersion.name
+      )}. ${returnText}`
     }
 
     if (!flag.silent) {
-      this.log(returnText)
-      return returnText
+      log(returnText, 'info')
     }
+
+    return returnText
   }
 }